+++ tests/inverted.rs
@@ -1,12 +1,12 @@
 use dummy_pin::LastStateDummyPin;
 use embedded_hal::digital::v2::{InputPin, OutputPin, StatefulOutputPin, ToggleableOutputPin};
 use embedded_hal_mock::pin::{Mock as PinMock, State, Transaction, TransactionKind};
-use inverted_pin::{InvertedInputPin, InvertedOutputPin};
+use inverted_pin::InvertedPin;
 
 #[test]
 fn can_create() {
     let mock_pin = PinMock::new(&[]);
-    let pin = InvertedOutputPin::new(mock_pin);
+    let pin = InvertedPin::new(mock_pin);
     let mut mock_pin = pin.destroy();
     mock_pin.done();
 }
@@ -17,7 +17,7 @@ fn output_sets_inverted_logic() {
         Transaction::new(TransactionKind::Set(State::Low)),
         Transaction::new(TransactionKind::Set(State::High)),
     ]);
-    let mut pin = InvertedOutputPin::new(mock_pin);
+    let mut pin = InvertedPin::new(mock_pin);
     pin.set_high().unwrap();
     pin.set_low().unwrap();
     let mut mock_pin = pin.destroy();
@@ -30,7 +30,7 @@ fn input_gets_inverted_logic() {
         Transaction::new(TransactionKind::Get(State::Low)),
         Transaction::new(TransactionKind::Get(State::High)),
     ]);
-    let pin = InvertedInputPin::new(mock_pin);
+    let pin = InvertedPin::new(mock_pin);
     assert!(pin.is_high().unwrap());
     assert!(pin.is_low().unwrap());
     let mut mock_pin = pin.destroy();
@@ -42,7 +42,7 @@ fn input_gets_inverted_logic() {
 #[test]
 fn stateful_output_is_set_high_gets_inverted_logic() {
     let mock_pin = LastStateDummyPin::new_high();
-    let mut pin = InvertedOutputPin::new(mock_pin);
+    let mut pin = InvertedPin::new(mock_pin);
     pin.set_high().unwrap();
     assert!(pin.is_set_high().unwrap());
     let mock_pin = pin.destroy();
@@ -52,7 +52,7 @@ fn stateful_output_is_set_high_gets_inverted_logic() {
 #[test]
 fn stateful_output_is_set_low_gets_inverted_logic() {
     let mock_pin = LastStateDummyPin::new_low();
-    let mut pin = InvertedOutputPin::new(mock_pin);
+    let mut pin = InvertedPin::new(mock_pin);
     pin.set_low().unwrap();
     assert!(pin.is_set_low().unwrap());
     let mock_pin = pin.destroy();
@@ -62,7 +62,7 @@ fn stateful_output_is_set_low_gets_inverted_logic() {
 #[test]
 fn output_can_toggle() {
     let mock_pin = LastStateDummyPin::new_high();
-    let mut pin = InvertedOutputPin::new(mock_pin);
+    let mut pin = InvertedPin::new(mock_pin);
     pin.toggle().unwrap();
     let mock_pin = pin.destroy();
     assert!(mock_pin.is_low().unwrap());