+++ CHANGELOG.md
@@ -7,9 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-02-18
+
+### Changed
+- [breaking-change] Merged `InvertedOutputPin` and `InvertedInputPin` together into
+  the `InvertedPin` struct.
+
 ## [0.1.0] - 2022-02-17
 
 This is the initial release to crates.io.
 
-[Unreleased]: https://github.com/eldruin/inverted-pin-rs/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/eldruin/inverted-pin-rs/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/eldruin/inverted-pin-rs/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/eldruin/inverted-pin-rs/releases/tag/v0.1.0